WIP: Add shading step to the deferred pipeline
Looks like the geometry step doesn't work,
the textures are just black.. Investigating
