+++ src/graphics/renderer/deferred.rs
@@ -1,6 +1,3 @@
-use std::{borrow::Cow, mem::size_of};
-use zerocopy::{AsBytes, FromBytes};
-
 //
 // GBuffer types
 //
@@ -102,7 +99,7 @@ impl GBuffer {
     }
 }
 
-fn create_texture(
+pub(super) fn create_texture(
     dimensions: (u32, u32),
     sample_count: u32,
     format: wgpu::TextureFormat,
@@ -185,28 +182,41 @@ impl<'a> DeferredContext<'a> {
 
     /// Shade the image drawn with deferred rendering
     /// and move on to rendering directly to the window.
-    pub fn shade(self) -> PostProcessContext<'a> {
+    pub fn shade(self, light: crate::DirectionalLight) -> PostShadeContext<'a> {
         let device = super::Renderer::device();
         let queue = super::Renderer::queue();
         queue.submit(Some(self.encoder.finish()));
 
         self.renderer.begin_frame();
-        let mut pp_ctx = PostProcessContext {
-            renderer: self.renderer,
-            encoder: Some(
-                device.create_command_encoder(&wgpu::CommandEncoderDescriptor::default()),
-            ),
-            is_first_draw: true,
-        };
 
         // run the shading
 
-        {
-            let mut shade_pass = pp_ctx.pass();
-            // TODO
-        }
+        let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor::default());
+        let mut shade_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
+            label: Some("shade"),
+            color_attachments: &[Some(wgpu::RenderPassColorAttachment {
+                // frame was just begun so this must exist
+                view: &self.renderer.active_frame.as_ref().unwrap().view,
+                resolve_target: None,
+                ops: wgpu::Operations {
+                    load: wgpu::LoadOp::Clear(wgpu::Color::BLACK),
+                    store: wgpu::StoreOp::Store,
+                },
+            })],
+            depth_stencil_attachment: None,
+            occlusion_query_set: None,
+            timestamp_writes: None,
+        });
 
-        pp_ctx
+        self.renderer
+            .deferred_shading_pl
+            .draw(&mut shade_pass, light);
+        drop(shade_pass);
+
+        PostShadeContext {
+            renderer: self.renderer,
+            encoder: Some(encoder),
+        }
     }
 }
 
@@ -215,237 +225,51 @@ pub struct DeferredPass<'a> {
     pub target_size: (u32, u32),
 }
 
-pub struct PostProcessContext<'a> {
+pub struct PostShadeContext<'a> {
     renderer: &'a mut super::Renderer,
     // encoder in an Option so that we can take it out
     // and submit it on drop without unsafe
     encoder: Option<wgpu::CommandEncoder>,
-    is_first_draw: bool,
 }
 
-impl<'a> PostProcessContext<'a> {
+impl<'a> PostShadeContext<'a> {
     /// Begin a render pass that draws on top of what's already in the window.
     pub fn pass(&mut self) -> wgpu::RenderPass {
         // encoder always exists, it's only removed on drop
         let encoder = self.encoder.as_mut().unwrap();
-        let pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-            label: Some("deferred"),
+        encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
+            label: Some("post-shade"),
             color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                // frame was begun in DeferredContext::submit,
-                // which is the only way to create a PostProcessContext
-                view: &self.renderer.active_frame.as_ref().unwrap().view,
-                resolve_target: None,
+                view: &self.renderer.msaa_view,
+                resolve_target: self.renderer.active_frame.as_ref().map(|f| &f.view),
                 ops: wgpu::Operations {
-                    load: if self.is_first_draw {
-                        wgpu::LoadOp::Clear(wgpu::Color::BLACK)
-                    } else {
-                        wgpu::LoadOp::Load
-                    },
+                    load: wgpu::LoadOp::Load,
                     store: wgpu::StoreOp::Store,
                 },
             })],
             depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
                 view: &self.renderer.gbufs.depth,
                 depth_ops: Some(wgpu::Operations {
-                    load: if self.is_first_draw {
-                        wgpu::LoadOp::Clear(0.)
-                    } else {
-                        wgpu::LoadOp::Load
-                    },
+                    load: wgpu::LoadOp::Load,
                     store: wgpu::StoreOp::Store,
                 }),
                 stencil_ops: None,
             }),
             occlusion_query_set: None,
             timestamp_writes: None,
-        });
-
-        self.is_first_draw = false;
-
-        pass
-    }
-}
-
-//
-// shading pipeline
-//
-
-pub struct ShadingPipeline {
-    pipeline: wgpu::RenderPipeline,
-    gbufs_bind_group_layout: wgpu::BindGroupLayout,
-    // bind group must be recreated when the window size changes
-    pub(super) gbufs_bind_group: wgpu::BindGroup,
-    light_buf: wgpu::Buffer,
-    light_bind_group: wgpu::BindGroup,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, AsBytes, FromBytes)]
-struct LightUniforms {
-    direct_color: [f32; 3],
-    _pad0: u32,
-    ambient_color: [f32; 3],
-    _pad1: u32,
-    direction: [f32; 3],
-    _pad2: u32,
-}
-
-impl From<crate::DirectionalLight> for LightUniforms {
-    fn from(l: crate::DirectionalLight) -> Self {
-        Self {
-            direct_color: l.direct_color,
-            _pad0: 0,
-            ambient_color: l.ambient_color,
-            _pad1: 0,
-            direction: l.direction.normalized().into(),
-            _pad2: 0,
-        }
-    }
-}
-
-impl ShadingPipeline {
-    pub fn new(gbufs: &GBuffers) -> Self {
-        let device = super::Renderer::device();
-
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: Some("deferred shading"),
-            source: wgpu::ShaderSource::Wgsl(Cow::Borrowed(include_str!(
-                "../shaders/deferred_shade.wgsl"
-            ))),
-        });
-
-        // gbuffer bind group
-
-        let tex_entry = |binding: u32| wgpu::BindGroupLayoutEntry {
-            binding,
-            visibility: wgpu::ShaderStages::FRAGMENT,
-            ty: wgpu::BindingType::Texture {
-                sample_type: wgpu::TextureSampleType::Float { filterable: false },
-                view_dimension: wgpu::TextureViewDimension::D2,
-                multisampled: false,
-            },
-            count: None,
-        };
-        let gbufs_bind_group_layout =
-            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-                label: Some("deferred shading gbuffers"),
-                entries: &[
-                    tex_entry(0),
-                    tex_entry(1),
-                    tex_entry(2),
-                    wgpu::BindGroupLayoutEntry {
-                        binding: 3,
-                        visibility: wgpu::ShaderStages::FRAGMENT,
-                        ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::NonFiltering),
-                        count: None,
-                    },
-                ],
-            });
-        let gbufs_bind_group = Self::create_gbufs_bind_group(&gbufs_bind_group_layout, gbufs);
-
-        // light
-
-        let light_buf = device.create_buffer(&wgpu::BufferDescriptor {
-            size: size_of::<LightUniforms>() as _,
-            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-            label: Some("mesh lights"),
-            mapped_at_creation: false,
-        });
-
-        let light_bind_group_layout =
-            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-                entries: &[wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: wgpu::BufferSize::new(size_of::<LightUniforms>() as _),
-                    },
-                    count: None,
-                }],
-                label: Some("deferred lights"),
-            });
-
-        let light_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("deferred lights"),
-            layout: &light_bind_group_layout,
-            entries: &[wgpu::BindGroupEntry {
-                binding: 0,
-                resource: light_buf.as_entire_binding(),
-            }],
-        });
-
-        // pipeline
-
-        let pl_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: Some("deferred shading"),
-            bind_group_layouts: &[&gbufs_bind_group_layout, &light_bind_group_layout],
-            push_constant_ranges: &[],
-        });
-        let pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("deferred shading"),
-            layout: Some(&pl_layout),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &[],
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                targets: &[Some(super::SWAPCHAIN_FORMAT.into())],
-            }),
-            primitive: wgpu::PrimitiveState {
-                topology: wgpu::PrimitiveTopology::TriangleList,
-                front_face: wgpu::FrontFace::Ccw,
-                cull_mode: None,
-                ..Default::default()
-            },
-            depth_stencil: None,
-            multisample: wgpu::MultisampleState::default(),
-            multiview: None,
-        });
-
-        Self {
-            pipeline,
-            gbufs_bind_group_layout,
-            gbufs_bind_group,
-            light_buf,
-            light_bind_group,
-        }
-    }
-
-    fn create_gbufs_bind_group(
-        layout: &wgpu::BindGroupLayout,
-        gbufs: &GBuffers,
-    ) -> wgpu::BindGroup {
-        let device = super::Renderer::device();
-        device.create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("deferred shading gbuffers"),
-            layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: wgpu::BindingResource::TextureView(&gbufs.position.view),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::TextureView(&gbufs.normal.view),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 2,
-                    resource: wgpu::BindingResource::TextureView(&gbufs.albedo.view),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 3,
-                    resource: wgpu::BindingResource::Sampler(&gbufs.sampler),
-                },
-            ],
         })
     }
 
-    pub(super) fn update_gbufs_bind_group(&mut self, gbufs: &GBuffers) {
-        self.gbufs_bind_group = Self::create_gbufs_bind_group(&self.gbufs_bind_group_layout, gbufs);
+    #[inline]
+    pub fn encoder_mut(&mut self) -> &mut wgpu::CommandEncoder {
+        self.encoder.as_mut().unwrap()
+    }
+}
+
+impl<'a> Drop for PostShadeContext<'a> {
+    // automatically submit on drop
+    fn drop(&mut self) {
+        let queue = super::Renderer::queue();
+        queue.submit(self.encoder.take().map(|enc| enc.finish()));
     }
 }
